# openapi/openapi.yml
openapi: 3.0.1
info:
  title: Каркас ППРБ
  description: Пример каркасного приложения ППРБ
  version: 01.00.000
servers:
  - url: /
tags:
  - name: draft
    description: Работа с черновиком
paths:
  /v1/draft:
    post:
      tags:
        - draft
      summary: Создание черновика
      description: Создание черновика
      operationId: createDraft
      requestBody:
        description: Запрос на создание черновика
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftCreateRequest'
        required: true
      responses:
        201:
          description: Успех
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftCreateResult'
        400:
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/draft/{uuid}:
    get:
      tags:
        - draft
      summary: Получение черновика по идентификатору
      description: Получение черновика по идентификатору
      operationId: getDraft
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            format: uuid
          required: true
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          description: Уникальный идентификатор черновика
      responses:
        200:
          description: Успех
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftView'
        400:
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - draft
      summary: Обновление черновика
      description: Обновление черновика
      operationId: updateDraft
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            format: uuid
          required: true
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          description: Уникальный идентификатор черновика
      requestBody:
        description: Запрос на обновление черновика
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftUpdateRequest'
        required: true
      responses:
        200:
          description: Успех
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftView'
        400:
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - draft
      summary: Удаление черновика
      description: Удаление черновика
      operationId: deleteDraft
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            format: uuid
          required: true
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          description: Уникальный идентификатор черновика
      responses:
        200:
          description: Успех
        400:
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/draft/list:
    get:
      tags:
        - draft
      summary: Получение списка черновиков по имени
      description: Получение списка черновиков по имени (по like)
      operationId: listDrafts
      parameters:
        - in: query
          name: name
          schema:
            type: string
          required: true
          example: "Наименование"
          description: Наименование черновика для поиска
        - in: query
          name: page
          schema:
            type: string
            format: int32
          example: "0"
          description: Номер страницы
        - in: query
          name: offset
          schema:
            type: string
            format: int32
          example: "0"
          description: Смещение пагинации
        - in: query
          name: size
          schema:
            type: string
            format: int32
          example: "10"
          description: Размер страницы
        - in: query
          name: sort
          schema:
            type: string
            format: int32
          example: "name,ASC"
          description: Сортировка записей
      responses:
        200:
          description: Успех
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'
        400:
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    DraftCreateRequest:
      required:
        - name
        - digitalId
      type: object
      properties:
        name:
          minLength: 1
          maxLength: 20
          type: string
          description: Наименование черновика
          example: "Наименование"
        digitalId:
          maxLength: 38
          pattern: ^\d+$
          type: string
          description: Идентификатор личного кабинета в фабрике Digital
          format: int64
          example: "1000000000000000"
        content:
          maxLength: 4000
          type: string
          description: Содержимое черновика
          example: "Содержимое черновика"
      description: Запрос на создание черновика
    DraftCreateResult:
      required:
        - guid
      type: object
      properties:
        guid:
          maxLength: 36
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
          type: string
          description: Уникальный идентификатор черновика
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    DraftUpdateRequest:
      type: object
      properties:
        content:
          maxLength: 4000
          type: string
          description: Содержимое черновика
          example: "Содержимое черновика"
      description: Запрос на создание черновика
    DraftView:
      required:
        - guid
        - digitalId
        - name
        - createDate
        - modifyDate
      type: object
      properties:
        guid:
          maxLength: 36
          pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
          type: string
          description: Уникальный идентификатор черновика
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        digitalId:
          maxLength: 38
          pattern: ^\d+$
          type: string
          description: Идентификатор личного кабинета в фабрике Digital
          format: int64
          example: "1000000000000000"
        name:
          minLength: 1
          maxLength: 20
          type: string
          description: Наименование черновика
          example: "Наименование"
        content:
          maxLength: 4000
          type: string
          description: Содержимое черновика
          example: "Содержимое черновика"
        createDate:
          type: string
          format: date-time
          description: Дата создания черновика
          example: "2021-08-12T14:24:47Z"
        modifyDate:
          type: string
          format: date-time
          description: Дата изменения черновика
          example: "2021-08-12T14:26:03Z"
      description: Успешный ответ
    ListResponse:
      description: specific content List of the Page response
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DraftView'
        page:
          description: Номер страницы
          type: integer
        size:
          description: Размер страницы
          type: integer
        totalElements:
          description: Общее количество элементов по запросу
          type: integer
        last:
          description: Признак последней страницы
          type: boolean
    ErrorResponse:
      required:
        - errorCode
        - errorDesc
        - errorName
      type: object
      properties:
        errorCode:
          maxLength: 10
          pattern: ^(.*)$
          type: string
          description: Код ошибки
          example: "400"
        errorName:
          maxLength: 255
          pattern: ^(.*)$
          type: string
          description: Наименование ошибки
          example: BAD_REQUEST
        errorDesc:
          maxLength: 1000
          pattern: ^(.*)$
          type: string
          description: Описание ошибки
          example: Длина поля content не должна превышать 4000 символов
      description: Неуспешный ответ